Created a new Boot Menu
This new boot menu has a proposed logo for SynthLog and a Dropdown menu that lacks functionality.

The "Create New Notebook" menu item will open a new webpage that asks them to name the new notebook. This will create a new folder with that name in the section of the application that stores folders.

The "open menu" folder item will open a file manager menu that starts at the location of the saved folders. The user will then select a folder and it will open a new webpage that has that folder loaded.

Please make sure to install Material UI on your developer side